Moved keycode and mouse button
